# docs/swagger.yaml
definitions:
...
  dokumen.PerdaPendirianModel:
    properties:
      file:
        type: string
      id:
        type: string
      id_bumd:
        type: string
      keterangan:
        type: string
      modal_dasar:
        type: number
      nomor:
        type: string
      tanggal:
        type: string
    type: object
...
  utils.DataValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email address
        type: string
    type: object
...
  utils.RequestError:
    properties:
      code:
        example: 422
        type: integer
      fields:
        items:
          $ref: '#/definitions/utils.DataValidationError'
        type: array
      message:
        example: Invalid email address
        type: string
    type: object
info:
  contact:
    email: lifelinejar@mail.com
    name: API Support
  description: BUMD Service Rest API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BUMD Service
  version: "1.0"
paths:
...
  /strict/bumd/{id_bumd}/perda_pendirian:
    get:
      description: get data perda pendirian.
      operationId: perda_pendirian-index
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Halaman yang ditampilkan
        in: query
        name: page
        type: integer
      - description: Jumlah data per halaman, maksimal 5 data per halaman
        in: query
        name: limit
        type: integer
      - description: Pencarian
        in: query
        name: search
        type: string
      - description: Modal Dasar Min
        format: float64
        in: query
        name: modal_dasar_min
        type: number
      - description: Modal Dasar Max
        format: float64
        in: query
        name: modal_dasar_max
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dokumen.PerdaPendirianModel'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data perda pendirian
      tags:
      - Perda Pendirian
    post:
      consumes:
      - multipart/form-data
      description: create data perda pendirian.
      operationId: perda_pendirian-create
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Nomor Perda
        in: formData
        name: nomor_perda
        required: true
        type: string
      - description: Tanggal Perda
        in: formData
        name: tanggal_perda
        required: true
        type: string
      - description: Keterangan
        in: formData
        name: keterangan
        required: true
        type: string
      - description: File
        in: formData
        name: file
        type: file
      - description: Modal Dasar
        in: formData
        name: modal_dasar
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: create data perda pendirian
      tags:
      - Perda Pendirian
  /strict/bumd/{id_bumd}/perda_pendirian/{id}:
    delete:
      consumes:
      - application/json
      description: delete data perda pendirian.
      operationId: perda_pendirian-delete
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk delete data perda pendirian
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: delete data perda pendirian
      tags:
      - Perda Pendirian
    get:
      description: get data perda pendirian by id.
      operationId: perda_pendirian-view
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk get data perda pendirian
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dokumen.PerdaPendirianModel'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data perda pendirian by id
      tags:
      - Perda Pendirian
    put:
      consumes:
      - multipart/form-data
      description: update data perda pendirian.
      operationId: perda_pendirian-update
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk update data perda pendirian
        in: path
        name: id
        required: true
        type: integer
      - description: Nomor Perda
        in: formData
        name: nomor_perda
        required: true
        type: string
      - description: Tanggal Perda
        in: formData
        name: tanggal_perda
        required: true
        type: string
      - description: Keterangan
        in: formData
        name: keterangan
        required: true
        type: string
      - description: File
        in: formData
        name: file
        type: file
      - description: Modal Dasar
        in: formData
        name: modal_dasar
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: update data perda pendirian
      tags:
      - Perda Pendirian
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"